move travisnpmpublish to npm scripts
