+++ Gruntfile.js
@@ -1,5 +1,4 @@
 const browserify = require('browserify');
-const childProcess = require('child_process');
 const electronPackager = require('electron-packager');
 const fs = require('fs');
 
@@ -158,39 +157,6 @@ module.exports = (grunt) => {
       done();
     });
   });
-
-  grunt.registerTask(
-    'travisnpmpublish',
-    'Automatically publish to NPM via travis and create git tag.',
-    function () {
-      const done = this.async();
-      if (
-        process.env.TRAVIS_BRANCH != 'master' ||
-        (process.env.TRAVIS_PULL_REQUEST && process.env.TRAVIS_PULL_REQUEST != 'false')
-      ) {
-        grunt.log.writeln('Skipping travis npm publish');
-        return done();
-      }
-      childProcess.exec('git rev-parse --short HEAD', (err, stdout, stderr) => {
-        const hash = stdout.trim();
-        const packageJson = JSON.parse(fs.readFileSync('package.json'));
-        const version = packageJson.version;
-        packageJson.version += `+${hash}`;
-        fs.writeFileSync('package.json', `${JSON.stringify(packageJson, null, 2)}\n`);
-        fs.writeFileSync('.npmrc', '//registry.npmjs.org/:_authToken=' + process.env.NPM_TOKEN);
-        childProcess.exec('npm publish', (err) => {
-          if (err) done(err);
-          else
-            childProcess.exec(
-              `git tag v${version} && git push -q https://${process.env.GITHUB_TOKEN}@github.com/FredrikNoren/ungit.git v${version}`,
-              (err) => {
-                done(err);
-              }
-            );
-        });
-      });
-    }
-  );
 
   grunt.registerTask('electronpublish', ['zip_directories:electron']);
